// apps/web/generated/api/models/createFreelancerProfileRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateFreelancerProfileRequest {
  /** @minLength 1 */
  display_name: string;
  headline?: string | null;
  bio?: string | null;
  country_code?: string | null;
  city?: string | null;
  timezone?: string | null;
}